+++ lib/features/auth/providers/user_context_provider.dart
@@ -1,9 +1,9 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
-import 'package:rlv2_flutter/features/auth/notifiers/auth_notifier.dart';
-import 'package:rlv2_flutter/features/auth/notifiers/user_context_notifier.dart';
 import 'package:rlv2_flutter/features/auth/providers/auth_provider.dart';
 import 'package:rlv2_flutter/features/auth/services/user_session_context_service.dart';
+import 'package:rlv2_flutter/features/auth/view_models/auth_view_model.dart';
+import 'package:rlv2_flutter/features/auth/view_models/user_context_view_model.dart';
 import 'package:rlv2_flutter/utils/app_logger.dart';
 
 // provider
